requirements.txt
kafka-python==2.3.0
requests==2.32.3

